refactor: Rename 'friendrequestroom' package to 'friend'
