+++ src/main/java/com/example/syncrowbackend/friend/entity/FriendRequest.java
@@ -1,7 +1,7 @@
-package com.example.syncrowbackend.friendrequestroom.entity;
+package com.example.syncrowbackend.friend.entity;
 
 import com.example.syncrowbackend.common.entity.BaseTimeEntity;
-import com.example.syncrowbackend.friendrequestroom.enums.FriendRequestStatus;
+import com.example.syncrowbackend.friend.enums.FriendRequestStatus;
 import com.example.syncrowbackend.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
