+++ src/main/java/com/example/syncrowbackend/friend/enums/FriendRequestStatus.java
@@ -1,4 +1,4 @@
-package com.example.syncrowbackend.friendrequestroom.enums;
+package com.example.syncrowbackend.friend.enums;
 
 public enum FriendRequestStatus {
     ACCEPTED,
