+++ src/main/java/com/example/syncrowbackend/friend/service/GroupService.java
@@ -1,0 +1,4 @@
+package com.example.syncrowbackend.friend.service;
+
+public interface GroupService {
+}
