+++ src/main/java/com/example/syncrowbackend/friend/service/GroupServiceImpl.java
@@ -1,0 +1,9 @@
+package com.example.syncrowbackend.friend.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class GroupServiceImpl implements GroupService{
+}
