+++ src/main/java/com/example/syncrowbackend/friend/service/PostServiceImpl.java
@@ -1,17 +1,17 @@
-package com.example.syncrowbackend.friendrequestroom.service;
+package com.example.syncrowbackend.friend.service;
 
 import com.example.syncrowbackend.common.exception.CustomException;
 import com.example.syncrowbackend.common.exception.ErrorCode;
-import com.example.syncrowbackend.friendrequestroom.dto.FriendRequestPostDto;
-import com.example.syncrowbackend.friendrequestroom.dto.PostDto;
-import com.example.syncrowbackend.friendrequestroom.dto.PostRequestDto;
-import com.example.syncrowbackend.friendrequestroom.entity.FriendRequest;
-import com.example.syncrowbackend.friendrequestroom.entity.Group;
-import com.example.syncrowbackend.friendrequestroom.entity.Post;
-import com.example.syncrowbackend.friendrequestroom.enums.FriendRequestStatus;
-import com.example.syncrowbackend.friendrequestroom.repository.FriendRequestRepository;
-import com.example.syncrowbackend.friendrequestroom.repository.GroupRepository;
-import com.example.syncrowbackend.friendrequestroom.repository.PostRepository;
+import com.example.syncrowbackend.friend.dto.FriendRequestPostDto;
+import com.example.syncrowbackend.friend.dto.PostDto;
+import com.example.syncrowbackend.friend.dto.PostRequestDto;
+import com.example.syncrowbackend.friend.entity.FriendRequest;
+import com.example.syncrowbackend.friend.entity.Group;
+import com.example.syncrowbackend.friend.entity.Post;
+import com.example.syncrowbackend.friend.enums.FriendRequestStatus;
+import com.example.syncrowbackend.friend.repository.FriendRequestRepository;
+import com.example.syncrowbackend.friend.repository.GroupRepository;
+import com.example.syncrowbackend.friend.repository.PostRepository;
 import com.example.syncrowbackend.user.entity.User;
 import com.example.syncrowbackend.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
